Added support for interpreting single numerical literals as boolean values.

+++ compiler/source/codegen/abstract_syntax_tree/functions/function_call_node.cpp
@@ -13,7 +13,7 @@ namespace channel {
 		return visitor.visit_function_call_node(*this, context);
 	}
 
-	void function_call_node::print(int depth, const std::wstring& prefix, bool is_last) {
+	void function_call_node::print(u64 depth, const std::wstring& prefix, bool is_last) {
 		print_value(depth, prefix, "function call",	is_last);
 		console::out
 			<< "'"
