+++ compiler/source/codegen/abstract_syntax_tree/keywords/flow_control/if_else_node.h
@@ -18,7 +18,7 @@ namespace channel {
 			const codegen_context& context
 		) override;
 
-		void print(int depth, const std::wstring& prefix, bool is_last) override;
+		void print(u64 depth, const std::wstring& prefix, bool is_last) override;
 
 		const std::vector<node_ptr>& get_condition_nodes();
 		const std::vector<std::vector<node_ptr>>& get_branch_nodes();
