+++ compiler/source/codegen/abstract_syntax_tree/keywords/types/bool_node.h
@@ -17,7 +17,7 @@ namespace channel {
 			const codegen_context& context
 		) override;
 
-		void print(int depth, const std::wstring& prefix, bool is_last) override;
+		void print(u64 depth, const std::wstring& prefix, bool is_last) override;
 
 		bool get_value() const;
 	private:
