+++ compiler/source/codegen/abstract_syntax_tree/keywords/types/numerical_literal_node.h
@@ -15,7 +15,7 @@ namespace channel {
 			const codegen_context& context
 		) override;
 
-		void print(int depth, const std::wstring& prefix, bool is_last) override;
+		void print(u64 depth, const std::wstring& prefix, bool is_last) override;
 
 		const std::string& get_value() const;
 		const type& get_preferred_type() const;
