+++ compiler/source/codegen/abstract_syntax_tree/variables/array/array_access_node.cpp
@@ -13,7 +13,7 @@ namespace channel {
 		return visitor.visit_array_access_node(*this, context);
 	}
 
-	void array_access_node::print(int depth, const std::wstring& prefix, bool is_last) {
+	void array_access_node::print(u64 depth, const std::wstring& prefix, bool is_last) {
 		print_value(depth, prefix, "array access", is_last);
 		console::out << "\n";
