+++ compiler/source/codegen/abstract_syntax_tree/variables/array/array_allocation_node.h
@@ -18,7 +18,7 @@ namespace channel {
 			const codegen_context& context
 		) override;
 
-		void print(int depth, const std::wstring& prefix, bool is_last) override;
+		void print(u64 depth, const std::wstring& prefix, bool is_last) override;
 
 		const type& get_array_element_type() const;
 		const node_ptr& get_array_element_count_node() const;
