+++ compiler/source/codegen/abstract_syntax_tree/variables/array/array_assignment_node.cpp
@@ -15,7 +15,7 @@ namespace channel {
 		return visitor.visit_array_assignment_node(*this, context);
 	}
 
-	void array_assignment_node::print(int depth, const std::wstring& prefix, bool is_last) {
+	void array_assignment_node::print(u64 depth, const std::wstring& prefix, bool is_last) {
 		print_value(depth, prefix, "array assignment", is_last);
 		console::out << "\n";
