+++ compiler/source/codegen/abstract_syntax_tree/variables/array/array_assignment_node.h
@@ -19,7 +19,7 @@ namespace channel {
             const codegen_context& context
         ) override;
 
-        void print(int depth, const std::wstring& prefix, bool is_last) override;
+        void print(u64 depth, const std::wstring& prefix, bool is_last) override;
 
         const node_ptr& get_array_base_node() const;
         const std::vector<node_ptr>& get_array_element_index_nodes() const;
