+++ compiler/source/codegen/abstract_syntax_tree/variables/declaration/local_declaration_node.cpp
@@ -17,7 +17,7 @@ namespace channel {
 		return visitor.visit_local_declaration_node(*this, context);
 	}
 
-	void local_declaration_node::print(int depth, const std::wstring& prefix, bool is_last) {
+	void local_declaration_node::print(u64 depth, const std::wstring& prefix, bool is_last) {
 		print_value(
 			depth, 
 			prefix,
