+++ compiler/source/codegen/abstract_syntax_tree/variables/variable_node.cpp
@@ -11,7 +11,7 @@ namespace channel {
 		return visitor.visit_variable_node(*this, context);
 	}
 
-	void variable_node::print(int depth, const std::wstring& prefix, bool is_last) {
+	void variable_node::print(u64 depth, const std::wstring& prefix, bool is_last) {
 		print_value(
 			depth,
 			prefix,
